# [Xamarin.Android.Build.Tasks] ResolveLibraryProjectImports temp files